requirements.txt
streamlit
azure-ai-textanalytics
openai
azure-core